fix(agents): classify missing workspace spawn failures
